Changed the parameters for FlashSeriesWithOptoStep
Changed the parameters here to reflect pilot data findings. Reduced Opto intensity options, added more visual flashes, and tightened up the trial:
I think what this means for the experiment is to have a 3 sec opto stim. 3+ seconds seems to show another round of hyperpolarization, which is muddying up the result to the light flash. So:
2 second PRE TIME
0 sec - start STIM TIME
1 sec - VF (vis flash) 1
3 sec - Opto On (until 6)
3.2 sec - VF2 (right after opto)
4.5 sec - VF3
6 sec - Opto Off
6.2 sec - VF4 (should start right after opto ends [big response[
8.2 sec - VF5
10.2 sec - VF6
12.2 sec - VF7
14 sec - end STIM TIME
2 sec TAIL TIME
-that’s 2 + 14 + 2, or 18-second trials. It’s also 13 seconds from opto off to the first VF (8-10 sec should be sufficient)
-18sec trials * 3 opto intensities * 36 trials each = 1944sec total, or 32.4 minutes which we should be good with given Carl's sub 40 (really 30) goal. Tom says it is fine